+++ Plutarch/Bool.hs
@@ -29,7 +29,7 @@ import Generics.SOP (
   ccompare_NS,
   hcliftA2,
  )
-import Plutarch.Internal (punsafeAsClosedTerm)
+import Plutarch.Internal (plet, punsafeAsClosedTerm)
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
 import Plutarch.Internal.Other (
   DerivePNewtype,
@@ -45,11 +45,12 @@ import Plutarch.Internal.Other (
   pmatch,
   pto,
   (#),
+  (#$),
   type (:-->),
  )
 import Plutarch.Lift (
   DerivePConstantDirect (DerivePConstantDirect),
-  PConstant,
+  PConstantDecl,
   PLifted,
   PUnsafeLiftDecl,
   pconstant,
@@ -62,7 +63,7 @@ data PBool (s :: S) = PTrue | PFalse
   deriving stock (Show)
 
 instance PUnsafeLiftDecl PBool where type PLifted PBool = Bool
-deriving via (DerivePConstantDirect Bool PBool) instance (PConstant Bool)
+deriving via (DerivePConstantDirect Bool PBool) instance PConstantDecl Bool
 
 instance PlutusType PBool where
   type PInner PBool _ = PBool
@@ -87,6 +88,13 @@ class PEq t => POrd t where
 
 infix 4 #<=
 infix 4 #<
+
+instance PEq PBool where
+  x #== y' = plet y' $ \y -> pif' # x # y #$ pnot # y
+
+instance POrd PBool where
+  x #< y = pif' # x # pconstant False # y
+  x #<= y = pif' # x # y # pconstant True
 
 instance PEq b => PEq (DerivePNewtype a b) where
   x #== y = pto x #== pto y