+++ Plutarch/DataRepr.hs
@@ -1,6 +1,6 @@
 module Plutarch.DataRepr (
   -- * DataRepr
-  I.PDataSum,
+  I.PDataSum (PDataSum),
   I.punDataSum,
   I.ptryIndexDataSum,
   I.DataReprHandlers (DRHNil, DRHCons),
@@ -15,12 +15,13 @@ module Plutarch.DataRepr (
   I.pdropDataRecord,
   I.DerivePConstantViaData (DerivePConstantViaData),
   I.pasDataSum,
+  I.PConstantData,
+  I.PLiftData,
 
   -- * Fields
   F.PDataFields (ptoFields, type PFields),
   F.pletFields,
   F.pfield,
-  F.hrecField,
   F.HRec,
 ) where