+++ plutarch-test/plutarch-base/Plutarch/ApiSpec.hs
@@ -1,6 +1,17 @@
-module Plutarch.ApiSpec (spec, ctx, info, purpose, validator, datum) where
+-- NOTE: This module also contains ScriptContext mocks, which should ideally
+-- moved to a module of its own after cleaning up to expose a easy to reason
+-- about API.
+module Plutarch.ApiSpec (
+  spec,
+  ctx,
+  validContext0,
+  validOutputs0,
+  invalidContext1,
+  d0Dat,
+  d0DatValue,
+  inp,
+) where
 
-import Test.Syd
 import Test.Tasty.HUnit
 
 import Control.Monad.Trans.Cont (cont, runCont)
@@ -8,7 +19,6 @@ import Plutus.V1.Ledger.Api
 import qualified Plutus.V1.Ledger.Interval as Interval
 import qualified Plutus.V1.Ledger.Value as Value
 
-import Plutarch (popaque)
 import Plutarch.Api.V1 (
   PCredential,
   PCurrencySymbol,
@@ -22,8 +32,7 @@ import Plutarch.Api.V1 (
 import Plutarch.Builtin (pasConstr, pforgetData)
 import Plutarch.Prelude
 import Plutarch.Test
-
--- import PlutusTx.AssocMap as PlutusMap
+import Test.Hspec
 
 spec :: Spec
 spec = do
@@ -167,11 +176,11 @@ getSym =
 checkSignatoryCont :: forall s. Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
 checkSignatoryCont = plam $ \ph ctx' ->
   pletFields @["txInfo", "purpose"] ctx' $ \ctx -> (`runCont` id) $ do
-    purpose <- cont (pmatch $ hrecField @"purpose" ctx)
+    purpose <- cont (pmatch $ getField @"purpose" ctx)
     pure $ case purpose of
       PSpending _ ->
         let signatories :: Term s (PBuiltinList (PAsData PPubKeyHash))
-            signatories = pfield @"signatories" # hrecField @"txInfo" ctx
+            signatories = pfield @"signatories" # getField @"txInfo" ctx
          in pif
               (pelem # pdata ph # signatories)
               -- Success!
@@ -185,9 +194,9 @@ checkSignatoryCont = plam $ \ph ctx' ->
 checkSignatoryTermCont :: Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
 checkSignatoryTermCont = plam $ \ph ctx' -> unTermCont $ do
   ctx <- tcont $ pletFields @["txInfo", "purpose"] ctx'
-  tcont (pmatch $ hrecField @"purpose" ctx) >>= \case
+  tcont (pmatch $ getField @"purpose" ctx) >>= \case
     PSpending _ -> do
-      let signatories = pfield @"signatories" # hrecField @"txInfo" ctx
+      let signatories = pfield @"signatories" # getField @"txInfo" ctx
       pure $
         pif
           (pelem # pdata ph # pfromData signatories)
@@ -203,3 +212,58 @@ getFields = phoistAcyclic $ plam $ \addr -> psndBuiltin #$ pasConstr # addr
 
 dummyCurrency :: CurrencySymbol
 dummyCurrency = Value.currencySymbol "\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11\x11"
+
+------------------- Mocking a ScriptContext ----------------------------------------
+
+validContext0 :: Term s PScriptContext
+validContext0 = mkCtx validOutputs0 validDatums1
+
+invalidContext1 :: Term s PScriptContext
+invalidContext1 = mkCtx invalidOutputs1 validDatums1
+
+mkCtx :: [TxOut] -> [(DatumHash, Datum)] -> Term s PScriptContext
+mkCtx outs l = pconstant (ScriptContext (info' outs l) purpose)
+  where
+    info' :: [TxOut] -> [(DatumHash, Datum)] -> TxInfo
+    info' outs dat =
+      info
+        { txInfoData = dat
+        , txInfoOutputs = outs
+        }
+
+validOutputs0 :: [TxOut]
+validOutputs0 =
+  [ TxOut
+      { txOutAddress =
+          Address (ScriptCredential validator) Nothing
+      , txOutValue = mempty
+      , txOutDatumHash = Just datum
+      }
+  ]
+
+invalidOutputs1 :: [TxOut]
+invalidOutputs1 =
+  [ TxOut
+      { txOutAddress =
+          Address (ScriptCredential validator) Nothing
+      , txOutValue = mempty
+      , txOutDatumHash = Just datum
+      }
+  , TxOut
+      { txOutAddress =
+          Address (ScriptCredential validator) Nothing
+      , txOutValue = mempty
+      , txOutDatumHash = Nothing
+      }
+  ]
+
+validDatums1 :: [(DatumHash, Datum)]
+validDatums1 =
+  [("d0", d0Dat)]
+
+-- | Mock datum that is a list of integers.
+d0Dat :: Datum
+d0Dat = Datum $ toBuiltinData d0DatValue
+
+d0DatValue :: [Integer]
+d0DatValue = [1 .. 10]