+++ plutarch-test/plutarch-base/Plutarch/ListSpec.hs
@@ -1,21 +1,19 @@
 module Plutarch.ListSpec (spec, integerList) where
 
-import Test.Syd
-import Test.Syd.Hedgehog ()
+import Data.List (find)
 
 import Plutarch.List (pconvertLists, pfoldl')
 import Plutarch.Prelude
-import Plutarch.Test
 
 import Hedgehog (Property)
-
-import Data.List (find)
-
-import Plutarch.Test.Property.Gen (genList, integerGen)
-import Plutarch.Test.Property.Util (haskPlutEquiv, viaBothPartial, viaPEq)
-
 import qualified Hedgehog.Gen as Gen
+import Hedgehog.Internal.Property (Property (propertyTest))
 import qualified Hedgehog.Range as Range
+import Plutarch.Test
+import Plutarch.Test.Property
+import Plutarch.Test.Property.Gen (genInteger, genList)
+import Test.Hspec (Spec, describe, it)
+import Test.Hspec.Hedgehog (hedgehog)
 
 integerList :: [Integer] -> Term s (PList PInteger)
 integerList xs = pconvertLists #$ pconstant @(PBuiltinList PInteger) xs
@@ -25,9 +23,9 @@ spec = do
   describe "list" $ do
     describe "properties" $ do
       describe "find" $ do
-        it "plutarch level find mirrors haskell level find" findTest
+        it "plutarch level find mirrors haskell level find" . hedgehog . propertyTest $ prop_pfindEquiv
       describe "elemAt" $ do
-        it "plutarch level elemAt mirrors haskell level elemAt" elemAtTest
+        it "plutarch level elemAt mirrors haskell level elemAt" . hedgehog . propertyTest $ prop_pelemAtEquiv
     plutarchDevFlagDescribe . pgoldenSpec $ do
       let xs10 :: Term _ (PList PInteger)
           xs10 = integerList [1 .. 10]
@@ -98,24 +96,30 @@ spec = do
               PFalse -> plet (phead # numList) $ \_x ->
                 ptail # numList
 
-findTest :: Property
-findTest =
-  haskPlutEquiv
-    viaPEq
+-- plutarch level find mirrors haskell level find
+prop_pfindEquiv :: Property
+prop_pfindEquiv =
+  prop_haskEquiv
+    @( 'OnPEq)
+    @( 'TotalFun)
     (find @[] @Integer even)
     (pfind # peven)
-    (genList integerGen)
+    (genList genInteger :* Nil)
   where
     peven :: Term s (PInteger :--> PBool)
     peven = plam $ \n -> pmod # n # 2 #== 0
 
-elemAtTest :: Property
-elemAtTest =
-  haskPlutEquiv
-    viaBothPartial
+-- plutarch level elemAt mirrors haskell level elemAt
+prop_pelemAtEquiv :: Property
+prop_pelemAtEquiv =
+  prop_haskEquiv
+    @( 'OnBoth)
+    @( 'PartialFun)
     elemAt
     pelemAt
-    (Gen.integral $ Range.linear (-10) 100, Gen.list (Range.linear 0 100) integerGen)
+    $ Gen.integral (Range.linear (-10) 100)
+      :* Gen.list (Range.linear 0 100) genInteger
+      :* Nil
 
 elemAt :: Integer -> [Integer] -> Integer
 elemAt n xs = xs !! fromInteger n